Split up ongoing election into components.
